requirements.txt
llama_cpp_python==0.3.8
numpy==2.2.5
gradio==5.27.0
